Move most caching methods to ActionDispatch::Caching, and let ActionMailer and ActionController to include it
